+++ common/src/main/java/org/somda/sdc/common/util/AnyDateTime.java
@@ -131,4 +131,29 @@ public class AnyDateTime {
          */
         void orElse(Consumer<T> consumer);
     }
+
+    /**
+     * Compares two {@linkplain AnyDateTime} objects on equality.
+     *
+     * @param object the right-hand side to compare against with this object.
+     * @return true if all attributes equal true, otherwise false.
+     */
+    @Override
+    public boolean equals(final Object object) {
+        if (this == object) return true;
+        if (object == null || getClass() != object.getClass()) return false;
+        final AnyDateTime that = (AnyDateTime) object;
+        return Objects.equals(local, that.local) &&
+                Objects.equals(offset, that.offset);
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hash(local, offset);
+    }
+
+    @Override
+    public String toString() {
+        return ObjectStringifier.stringify(this);
+    }
 }